% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_operations.R
\name{create_stream}
\alias{create_stream}
\title{Creates a new Kinesis video stream}
\usage{
create_stream(StreamName, DeviceName = NULL, MediaType = NULL,
  KmsKeyId = NULL, DataRetentionInHours = NULL)
}
\arguments{
\item{StreamName}{A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.}

\item{DeviceName}{The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.}

\item{MediaType}{The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see \href{http://www.iana.org/assignments/media-types/media-types.xhtml}{Media Types}. If you choose to specify the \code{MediaType}, see \href{https://tools.ietf.org/html/rfc6838#section-4.2}{Naming Requirements} for guidelines.

To play video on the console, the media must be H.264 encoded, and you need to specify this video type in this parameter as \code{video/h264}.

This parameter is optional; the default value is \code{null} (or empty in JSON).}

\item{KmsKeyId}{The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (\code{aws/kinesisvideo}) is used.

For more information, see \href{http://docs.aws.amazon.com/kms/latest/APIReference/API_DescribeKey.html#API_DescribeKey_RequestParameters}{DescribeKey}.}

\item{DataRetentionInHours}{The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the \code{DataRetentionInHours} value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.}
}
\description{
Creates a new Kinesis video stream.
}
\details{
When you create a new stream, Kinesis Video Streams assigns it a version number. When you change the stream's metadata, Kinesis Video Streams updates the version.

\code{CreateStream} is an asynchronous operation.

For information about how the service works, see \href{http://docs.aws.amazon.com/kinesisvideostreams/latest/dg/how-it-works.html}{How it Works}.

You must have permissions for the \code{KinesisVideo:CreateStream} action.
}
\examples{

}
