% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_operations.R
\name{describe_stream}
\alias{describe_stream}
\title{Returns the most current information about the specified stream}
\usage{
describe_stream(StreamName = NULL, StreamARN = NULL)
}
\arguments{
\item{StreamName}{The name of the stream.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream.}
}
\description{
Returns the most current information about the specified stream. You must specify either the \code{StreamName} or the \code{StreamARN}.
}
\examples{

}
