% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_operations.R
\name{get_data_endpoint}
\alias{get_data_endpoint}
\title{Gets an endpoint for a specified stream for either reading or writing}
\usage{
get_data_endpoint(APIName, StreamName = NULL, StreamARN = NULL)
}
\arguments{
\item{APIName}{The name of the API action for which to get an endpoint.}

\item{StreamName}{The name of the stream that you want to get the endpoint for. You must specify either this parameter or a \code{StreamARN} in the request.}

\item{StreamARN}{The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a \code{StreamName} in the request.}
}
\description{
Gets an endpoint for a specified stream for either reading or writing. Use this endpoint in your application to read from the specified stream (using the \code{GetMedia} or \code{GetMediaForFragmentList} operations) or write to it (using the \code{PutMedia} operation).
}
\details{
The returned endpoint does not have the API name appended. The client needs to add the API name to the returned endpoint.

In the request, specify the stream either by \code{StreamName} or \code{StreamARN}.
}
\examples{

}
