% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_operations.R
\name{list_streams}
\alias{list_streams}
\title{Returns an array of \code{StreamInfo} objects}
\usage{
list_streams(MaxResults = NULL, NextToken = NULL,
  StreamNameCondition = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of streams to return in the response. The default is 10,000.}

\item{NextToken}{If you specify this parameter, when the result of a \code{ListStreams} operation is truncated, the call returns the \code{NextToken} in the response. To get another batch of streams, provide this token in your next request.}

\item{StreamNameCondition}{Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.}
}
\description{
Returns an array of \code{StreamInfo} objects. Each object describes a stream. To retrieve only streams that satisfy a specific condition, you can specify a \code{StreamNameCondition}.
}
\examples{

}
