% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_package.R
\docType{package}
\name{paws.kinesisvideo-package}
\alias{paws.kinesisvideo}
\alias{paws.kinesisvideo-package}
\title{paws.kinesisvideo: Amazon Kinesis Video Streams}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
