% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideo_operations.R
\name{untag_stream}
\alias{untag_stream}
\title{Removes one or more tags from a stream}
\usage{
untag_stream(StreamARN = NULL, StreamName = NULL, TagKeyList)
}
\arguments{
\item{StreamARN}{The Amazon Resource Name (ARN) of the stream that you want to remove tags from.}

\item{StreamName}{The name of the stream that you want to remove tags from.}

\item{TagKeyList}{[required] A list of the keys of the tags that you want to remove.}
}
\description{
Removes one or more tags from a stream. In the request, specify only a tag key or keys; don't specify the value. If you specify a tag key that does not exist, it's ignored.
}
\details{
In the request, you must provide the \code{StreamName} or \code{StreamARN}.
}
\section{Accepted Parameters}{
\preformatted{untag_stream(
  StreamARN = "string",
  StreamName = "string",
  TagKeyList = list(
    "string"
  )
)
}
}

